Level End + Level 4 + 5 Scene
Now there's a level end script and object. 
The level end Object takes an object that contains all of the next level and it's Spawn. When both players are inside the trigger it turns off the old Level and turn on the Object that is on the reference. 
The script also teleport the players to the spawn point.